Refactor frontend and fix minor bugs
